ui: pack cell styles in 8 bytes; cleanup handling of unset colors
The VT100 and Curses backends unnecessarily had different storage
formats for a Cell's styling information. While technically
speaking Curses needs less space it was using more (more than
double what it needed). Instead both formats use an 8 byte
packing, 3 bytes for FG, 3 bytes for BG, 1 byte for attributes,
and 1 byte for additional property information for Vis to use.

Furthermore style definitions in vis are allowed to be empty (both
the FG and BG separately). This allows for the stuff like the
cursor to change the BG but to keep the FG (for example). This was
being handled through this cryptic logic which would try to check
if a color was "default" in which case it would get replaced by
whatever was already contained in the cell. Sometimes this
behaviour was not desirable (because it wasn't very robust) so it
was made configurable with a similarly cryptic "keep_non_default"
boolean that would get passed around whenever a style was set.
Everytime I would try to resolve issues surrounding styling that
boolean would result in unexepected behaviour and thwart my plans.

That boolean is no more. When a style is specified without color
information it gets flagged as such (using the properties field).
When that style is to be applied to a cell it will inherit the
cells existing colors. Cells are not allowed to have unset FG/BG.
They all start out with STYLE_DEFAULT. STYLE_DEFAULT uses the
terminal's defaults if the user specifies it without color
information. It is no longer possible to have any partially
defined styles trying to get pushed out to the terminal. There is
also a new `keep_attribute` keyword for keeping the underlying
cell's attributes in addition to the ones specified by the style
(see CURSORLINE changes for example).

This fixes many issues I know about with regards to styling in
both backends. This includes weird behaviour with the cursor and
cells with the DIM attribute (#1157, and my comment on #1356),
styles not being properly set when switching between themes, and
styles getting incorrectly transfered to the cursorline (#1277).
Indexed colors are now properly passed through in the VT100
backend (see #1083, though that is more complaining about lack of
documentation).

fixes #1157
fixes #1277

+++ util.h
@@ -134,7 +134,6 @@ static inline bool addu(size_t a, size_t b, size_t *c) {
 #define PATH_MAX 4096
 #endif
 
-
 typedef size_t    uz;
 typedef int64_t   s64;
 typedef uint64_t  u64;
@@ -149,11 +148,11 @@ typedef char      c8;
 #define VisDACount s32
 
 typedef struct {
-	ptrdiff_t  length;
-	uint8_t   *data;
+	s64  length;
+	u8  *data;
 } str8;
-#define str8(s)      (str8){.length = sizeof(s) - 1, .data = (uint8_t *)s}
-#define str8_comp(s)       {sizeof(s) - 1, (uint8_t *)s}
+#define str8(s)      (str8){.length = sizeof(s) - 1, .data = (u8 *)s}
+#define str8_comp(s)       {sizeof(s) - 1, (u8 *)s}
 
 typedef struct {
 	str8       *data;
